% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_width_summary}
\alias{bootstrap_width_summary}
\title{Compute Width of Bootstrap Distribution}
\usage{
bootstrap_width_summary(bootstrap_results, effect = "NIE")
}
\arguments{
\item{bootstrap_results}{The \code{bootstrap_results} property of a \code{\link{bound_ne}} fit with
\code{bootstrap = TRUE}, or the list returned by \code{compute_bootstrap_ci()}}

\item{effect}{Character string: "NIE" or "NDE"}
}
\value{
List with width statistics
}
\description{
Compute the mean, median, and range of bound widths from bootstrap samples.
}
