% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{compute_bound_se}
\alias{compute_bound_se}
\title{Compute Standard Errors for Bounds}
\usage{
compute_bound_se(bootstrap_results)
}
\arguments{
\item{bootstrap_results}{The \code{bootstrap_results} property of a \code{\link{bound_ne}} fit with
\code{bootstrap = TRUE}, or the list returned by \code{compute_bootstrap_ci()}}
}
\value{
Named vector of standard errors
}
\description{
Estimate standard errors for the lower and upper bounds using the
bootstrap distribution.
}
