% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_helpers.R
\name{compute_naive_effects}
\alias{compute_naive_effects}
\title{Compute Naive Effects}
\usage{
compute_naive_effects(
  data,
  exposure,
  mediator,
  outcome,
  confounders,
  effect_scale
)
}
\description{
Compute Naive Effects
}
\keyword{internal}
