% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_helpers.R
\name{compute_proportion_mediated}
\alias{compute_proportion_mediated}
\title{Compute Proportion Mediated}
\usage{
compute_proportion_mediated(nie, nde, effect_scale = "OR")
}
\arguments{
\item{nie}{Natural Indirect Effect}

\item{nde}{Natural Direct Effect}

\item{effect_scale}{Character string: "OR", "RR", or "RD"}
}
\value{
Proportion mediated (PM)
}
\description{
Compute the proportion of the total effect that is mediated
}
\keyword{internal}
