% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_helpers.R
\name{compute_relative_precision}
\alias{compute_relative_precision}
\title{Compute Relative Precision}
\usage{
compute_relative_precision(bounds_width, naive_ci_width)
}
\arguments{
\item{bounds_width}{Width of partial identification bounds}

\item{naive_ci_width}{Width of naive 95% CI}
}
\value{
Numeric: ratio of naive CI width to bounds width
}
\description{
Compute the relative precision of bounds compared to a naive estimate.
}
\keyword{internal}
