% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-classes.R
\name{.sensitivity_region_class}
\alias{.sensitivity_region_class}
\title{Sensitivity Region Class}
\usage{
.sensitivity_region_class(
  sn0_range = integer(0),
  sp0_range = integer(0),
  psi_sn_range = integer(0),
  psi_sp_range = integer(0)
)
}
\description{
Defines the sensitivity parameter space Theta_Psi for differential
misclassification analysis.
}
\keyword{internal}
