% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_helpers.R
\name{get_default_sensitivity_region}
\alias{get_default_sensitivity_region}
\title{Get Default Sensitivity Region}
\usage{
get_default_sensitivity_region()
}
\description{
Get Default Sensitivity Region
}
\keyword{internal}
