% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falsification_summary.R
\name{print.falsification_summary}
\alias{print.falsification_summary}
\title{Print Method for falsification_summary}
\usage{
\method{print}{falsification_summary}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{falsification_summary}}

\item{digits}{Integer. Number of decimal places. Default is 3.}

\item{...}{Additional arguments (currently unused)}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for falsification_summary
}
