% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-classes.R
\name{s7-classes}
\alias{s7-classes}
\title{S7 Class Definitions for medrobust Package}
\description{
This file contains all S7 class definitions for the medrobust package,
replacing the previous S3 implementation with a modern, type-safe OOP system.
}
\keyword{internal}
