% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_helpers.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate Inputs}
\usage{
validate_inputs(
  data,
  exposure,
  mediator,
  outcome,
  confounders,
  misclassified_variable,
  sensitivity_region,
  n_grid
)
}
\description{
Validate Inputs
}
\keyword{internal}
