% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_helpers.R
\name{validate_sensitivity_region}
\alias{validate_sensitivity_region}
\title{Validate Sensitivity Region}
\usage{
validate_sensitivity_region(sens_region)
}
\description{
Validate Sensitivity Region
}
\keyword{internal}
