% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{as.list.sensitivity_region}
\alias{as.list.sensitivity_region}
\title{Convert sensitivity_region S7 object to list}
\usage{
\method{as.list}{sensitivity_region}(x, ...)
}
\arguments{
\item{x}{A sensitivity_region S7 object}

\item{...}{Additional arguments (ignored)}
}
\value{
A list with sn0_range, sp0_range, psi_sn_range, psi_sp_range
}
\description{
Convert sensitivity_region S7 object to list
}
