% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-classes.R
\name{as.list.sensitivity_region}
\alias{as.list.sensitivity_region}
\title{Convert sensitivity_region to list}
\usage{
\method{as.list}{sensitivity_region}(x, ...)
}
\arguments{
\item{x}{sensitivity_region S7 object}

\item{...}{Additional arguments (ignored)}
}
\value{
Named list
}
\description{
Convert sensitivity_region to list
}
