% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-classes.R
\name{as_sensitivity_region}
\alias{as_sensitivity_region}
\title{Create sensitivity_region object from list}
\usage{
as_sensitivity_region(region_list)
}
\arguments{
\item{region_list}{List with sn0_range, sp0_range, psi_sn_range, psi_sp_range}
}
\value{
sensitivity_region S7 object
}
\description{
Create sensitivity_region object from list
}
