% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{compute_bound_se}
\alias{compute_bound_se}
\title{Compute Standard Errors for Bounds}
\usage{
compute_bound_se(bootstrap_results)
}
\arguments{
\item{bootstrap_results}{List returned by compute_bootstrap_ci}
}
\value{
Named vector of standard errors
}
\description{
Estimate standard errors for the lower and upper bounds using the
bootstrap distribution.
}
