% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falsification_summary.R
\name{extract_bounds}
\alias{extract_bounds}
\title{Extract Compatible Parameter Sets}
\usage{
extract_bounds(bounds_object)
}
\arguments{
\item{bounds_object}{An object of class \code{medrobust_bounds}}
}
\value{
A data frame containing compatible parameter sets
}
\description{
Extract the compatible parameter sets from a bounds analysis.
}
