% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falsification_summary.R
\name{extract_falsified_region}
\alias{extract_falsified_region}
\title{Extract Falsified Region}
\usage{
extract_falsified_region(bounds_object)
}
\arguments{
\item{bounds_object}{An object of class \code{medrobust_bounds}}
}
\value{
A data frame containing parameter sets that were falsified
}
\description{
Extract the subset of the sensitivity region that is empirically falsified.
}
