% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-classes.R
\name{medrobust_bounds}
\alias{medrobust_bounds}
\title{Medrobust Bounds Class}
\usage{
medrobust_bounds(
  NIE_lower = integer(0),
  NIE_upper = integer(0),
  NDE_lower = integer(0),
  NDE_upper = integer(0),
  compatible_sets = (function (.data = list(), row.names = NULL) 
 {
     if
    (is.null(row.names)) {
         list2DF(.data)
     }
     else {
         out <-
    list2DF(.data, length(row.names))
         attr(out, "row.names") <- row.names
      
      out
     }
 })(),
  n_compatible = integer(0),
  n_evaluated = integer(0),
  falsified_proportion = integer(0),
  effect_scale = character(0),
  misclassified_variable = character(0),
  sensitivity_region = sensitivity_region(),
  naive_estimates = list(),
  bootstrap_results = NULL,
  data_summary = list(),
  call = NULL
)
}
\description{
S7 class for storing partial identification bounds for natural direct
and indirect effects under differential misclassification.
}
