% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-classes.R
\name{medrobust_bounds}
\alias{medrobust_bounds}
\title{Medrobust Bounds Class}
\arguments{
\item{NIE_lower}{Lower bound for Natural Indirect Effect}

\item{NIE_upper}{Upper bound for Natural Indirect Effect}

\item{NDE_lower}{Lower bound for Natural Direct Effect}

\item{NDE_upper}{Upper bound for Natural Direct Effect}

\item{compatible_sets}{Data frame of parameter sets compatible with data}

\item{n_compatible}{Number of compatible parameter sets}

\item{n_evaluated}{Total number of parameter sets evaluated}

\item{falsified_proportion}{Proportion of parameter space falsified}

\item{effect_scale}{Character: "OR", "RR", or "RD"}

\item{misclassified_variable}{Character: "exposure" or "mediator"}

\item{sensitivity_region}{Sensitivity region specification}

\item{naive_estimates}{List of naive effect estimates}

\item{bootstrap_results}{Bootstrap inference results (if computed)}

\item{data_summary}{Summary statistics from the data}

\item{call}{The function call that created this object}
}
\description{
S7 class for storing partial identification bounds for natural direct
and indirect effects under differential misclassification.
}
