% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-classes.R
\name{power_analysis_result}
\alias{power_analysis_result}
\title{Power Analysis Result Class}
\usage{
power_analysis_result(
  power_curve = (function (.data = list(), row.names = NULL) 
 {
     if
    (is.null(row.names)) {
         list2DF(.data)
     }
     else {
         out <-
    list2DF(.data, length(row.names))
         attr(out, "row.names") <- row.names
      
      out
     }
 })(),
  true_effect = integer(0),
  target_power = integer(0),
  target_width = integer(0),
  recommended_n_power = NA_integer_,
  recommended_n_width = NA_integer_,
  simulation_params = list()
)
}
\description{
S7 class for storing power analysis results for partial identification bounds.
}
