% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-classes.R
\name{power_analysis_result}
\alias{power_analysis_result}
\title{Power Analysis Result Class}
\arguments{
\item{power_curve}{Data frame with power and width by sample size}

\item{true_effect}{True effect value used in simulations}

\item{target_power}{Target power level for sample size recommendations}

\item{target_width}{Target bound width for sample size recommendations}

\item{recommended_n_power}{Recommended sample size to achieve target power}

\item{recommended_n_width}{Recommended sample size to achieve target width}

\item{simulation_params}{List of simulation parameters used}
}
\description{
S7 class for storing power analysis results for partial identification bounds.
}
