% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_compatibility.R
\name{print.compatibility_test}
\alias{print.compatibility_test}
\title{Print Method for compatibility_test}
\usage{
\method{print}{compatibility_test}(x, ...)
}
\arguments{
\item{x}{An object of class \code{compatibility_test}}

\item{...}{Additional arguments (currently unused)}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for compatibility_test
}
