% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7-classes.R
\name{simulated_dm_data}
\alias{simulated_dm_data}
\title{Simulated Data with Differential Misclassification Class}
\usage{
simulated_dm_data(
  observed = (function (.data = list(), row.names = NULL) 
 {
     if
    (is.null(row.names)) {
         list2DF(.data)
     }
     else {
         out <-
    list2DF(.data, length(row.names))
         attr(out, "row.names") <- row.names
      
      out
     }
 })(),
  truth = NULL,
  true_effects = list(),
  generation_params = list(),
  misclassification_applied = list()
)
}
\description{
S7 class for storing simulated data with known differential misclassification,
used for power analysis and methods validation.
}
