% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_compatibility.R
\name{summary.compatibility_test}
\alias{summary.compatibility_test}
\title{Summary Method for compatibility_test}
\usage{
\method{summary}{compatibility_test}(object, ...)
}
\arguments{
\item{object}{An object of class \code{compatibility_test}}

\item{...}{Additional arguments (currently unused)}
}
\value{
Invisibly returns the input object
}
\description{
Summary Method for compatibility_test
}
