% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falsification_summary.R
\name{test_multiple_hypotheses}
\alias{test_multiple_hypotheses}
\title{Test Multiple Hypotheses}
\usage{
test_multiple_hypotheses(
  data,
  exposure,
  mediator,
  outcome,
  confounders,
  psi_list,
  misclassified_variable
)
}
\arguments{
\item{data}{Data frame}

\item{exposure}{Character string}

\item{mediator}{Character string}

\item{outcome}{Character string}

\item{confounders}{Character vector}

\item{psi_list}{List of parameter sets to test}

\item{misclassified_variable}{Character string}
}
\value{
Data frame with test results for each hypothesis
}
\description{
Test multiple hypotheses about misclassification parameters simultaneously.
}
